Move components to new directory
