+++ socialtours/src/components/EventComponents/Panel.js
